+++ qiskit/result/result.py
@@ -150,64 +150,6 @@ class Result(BaseModel):
         except KeyError:
             raise QISKitError('No unitary for circuit "{0}"'.format(circuit))
 
-    def get_snapshots(self, circuit=None):
-        """Get snapshots recorded during the run of an experiment.
-
-        Args:
-            circuit (str or QuantumCircuit or int or None): the index of the
-                experiment, as specified by ``data()``.
-
-        Returns:
-            dict[slot: dict[str: array]]: dictionary where the keys are the
-                requested snapshot slots, and the values are a dictionary of
-                the snapshots.
-
-        Raises:
-            QISKitError: if there are no snapshots for the experiment.
-        """
-        try:
-            return self._get_experiment(circuit).data.snapshots.to_dict()
-        except KeyError:
-            raise QISKitError('No snapshots for circuit "{0}"'.format(circuit))
-
-    def get_snapshot(self, slot=None, circuit=None):
-        """Get snapshot at a specific slot of an experiment.
-
-        Args:
-            slot (str): snapshot slot to retrieve. If None and there is only one
-                slot, return that one.
-            circuit (str or QuantumCircuit or int or None): the index of the
-                experiment, as specified by ``data()``.
-
-        Returns:
-            dict[slot: dict[str: array]]: list of 2^n_qubits complex amplitudes.
-
-        Raises:
-            QISKitError: if there is no snapshot at all, or in this slot
-        """
-        try:
-            snapshots_dict = self.get_snapshots(circuit)
-
-            if slot is None:
-                slots = list(snapshots_dict.keys())
-                if len(slots) == 1:
-                    slot = slots[0]
-                else:
-                    raise QISKitError("You have to select a slot when there "
-                                      "is more than one available")
-            snapshot_dict = snapshots_dict[slot]
-
-            snapshot_types = list(snapshot_dict.keys())
-            if len(snapshot_types) == 1:
-                snapshot_list = snapshot_dict[snapshot_types[0]]
-                if len(snapshot_list) == 1:
-                    return snapshot_list[0]
-                return snapshot_list
-            return snapshot_dict
-        except KeyError:
-            raise QISKitError('No snapshot at slot {0} for '
-                              'circuit "{1}"'.format(slot, circuit))
-
     def _get_experiment(self, key=None):
         """Return an experiment from a given key.