+++ test/python/circuit/test_circuit_operations.py
@@ -15,7 +15,7 @@ from qiskit import execute
 from qiskit import QISKitError
 from qiskit.quantum_info import state_fidelity
 
-from ..common import QiskitTestCase, bin_to_hex_keys
+from ..common import QiskitTestCase, bin_to_hex_keys, requires_cpp_simulator
 
 
 class TestCircuitOperations(QiskitTestCase):
@@ -73,6 +73,7 @@ class TestCircuitOperations(QiskitTestCase):
         self.assertRaises(QISKitError, qc1.__add__, qc2)
         self.assertRaises(QISKitError, qc1.__add__, qcr3)
 
+    @requires_cpp_simulator
     def test_combine_circuit_extension_instructions(self):
         """Test combining circuits containing barrier, initializer, snapshot
         """
@@ -86,10 +87,10 @@ class TestCircuitOperations(QiskitTestCase):
         qc2.snapshot(slot='1')
         qc2.measure(qr, cr)
         new_circuit = qc1 + qc2
-        backend = Aer.get_backend('qasm_simulator_py')
+        backend = Aer.get_backend('qasm_simulator')
         shots = 1024
         result = execute(new_circuit, backend=backend, shots=shots, seed=78).result()
-        snapshot_vectors = result.get_snapshot()
+        snapshot_vectors = result.data(0)['snapshots']['1']['statevector']
         fidelity = state_fidelity(snapshot_vectors[0], desired_vector)
         self.assertGreater(fidelity, 0.99)
 
@@ -150,6 +151,7 @@ class TestCircuitOperations(QiskitTestCase):
         self.assertRaises(QISKitError, qc1.__iadd__, qc2)
         self.assertRaises(QISKitError, qc1.__iadd__, qcr3)
 
+    @requires_cpp_simulator
     def test_extend_circuit_extension_instructions(self):
         """Test extending circuits containing barrier, initializer, snapshot
         """
@@ -167,7 +169,7 @@ class TestCircuitOperations(QiskitTestCase):
         shots = 1024
         result = execute(qc1, backend=backend, shots=shots, seed=78).result()
 
-        snapshot_vectors = result.get_snapshot('1')
+        snapshot_vectors = result.data(0)['snapshots']['1']['statevector']
         fidelity = state_fidelity(snapshot_vectors[0], desired_vector)
         self.assertGreater(fidelity, 0.99)